esbuild-plugin: move check.import handler into size-limit helper
