+++ packages/esbuild/get-config.js
@@ -1,21 +1,7 @@
-let { writeFile } = require('fs').promises
-let { join } = require('path')
-let mkdirp = require('mkdirp')
+let updateCheckWithImports = require('size-limit/updateCheckWithImports')
 
 module.exports = async function getConfig(limitConfig, check, output) {
-  if (check.import) {
-    let loader = ''
-    for (let i in check.import) {
-      let list = check.import[i].replace(/}|{/g, '').trim()
-      loader +=
-        `import ${check.import[i]} from ${JSON.stringify(i)}\n` +
-        `console.log(${list})\n`
-    }
-    await mkdirp(output)
-    let entry = join(output, 'index.js')
-    await writeFile(entry, loader)
-    check.files = entry
-  }
+  await updateCheckWithImports(check, output)
 
   let config = {
     entryPoints: Array.isArray(check.files) ? check.files : [check.files],
